Change test data to full name #94
- A bug existed where FULL_NAME was used in DataGateway but DataGateway
API used FULLNAME as a key for a user's full name. This didn't pick up
the full name from the API so always defaulted to the user's name, even
if full name was available. This happened on the database backend of the
API. Now DataGateway is using the ICAT backend, this bug no longer
exists, hence the test data has been changed