Replicating scigateway homepage with datagateway homepage as new separate plugin #431
This commit contains a lot of unfinished code in the wrong place that is being pushed for backup purposes.  Currently, the new homepage for datagateway is being kept as a new plugin as it didn't make sense to keep it anywhere else. Plugin does not currently load due to compiler errors. Fair amount of the content is already there, it is just in the wrong place and not linked up properly for the site.

Also duplicated code exists. This is because new code originally went
into dg-common but this is almost certainly the wrong place to put it.
Duplications will be removed in a later commit.