Moving homepage back to common, no longer its own package #431
Homepage component has now been moved back to common and is no longer its own separate package. Although this made sense for isolating the code, this is almost certainly too much overhead and unnecessary. Homepage is designed to be imported from common by another package, which then handles displaying it under certain circumstances.

Where the homepage is being stored and how it is being imported into scigateway is still up for debate but will be discussed more soon. Possible plan right now is to add an extra variable in datagateway-dataview-settings.json like "homepage": true, which will cause a redirect of any traffic to / to go to the datagateway homepage instead, which will possibly be a new route registered in that file. Whether this makes sense in the long term, e.g. lots of plugins being imported, is still unclear.

In this current commit, the HomePage component is displayed in all DataView pages purely for testing purposes.