Loads the homepage if the current location equals paths.homepage #431
The homepage URL is set in the paths variable. If the location equals this URL, render the homepage, otherwise render the usual dataview page. This may not be the best way of doing things; later on it may be preferable to use /datagateway as a root for all other datagateway pages, in which case all URLs would have to use this as a base.