Adding test for homepage displaying in dataview #431
This tests to see if the homepage component is rendered if the current location equals the set homepage location.

Had to take out the theming I had added back in from the last commit into the homepage. One line is still giving me grief. I'll have a proper look at it again later