Moving homepage content from contents.tsx to dataview-settings.json #431
Homepage text content now lives in the dataview settings JSON file. This makes the code more complex again and may be awkward with future plugins wanting to implement a homepage but this is how it is for now. A "homepage" field can now optionally be added to dataview-settings.json containing fields for "title", "explore-label", etc.  If these aren't specified, they are replaced with their label text instead.