Fetching facility cycles and storing in local state #658
More of a checkpoint than a real commit but contains mostly functioning code with TODOs.

This currently has all the code for making an axios request to {{datagateway-api }}/facilitycycles in the component file, which won't be the case in the long term but is for testing purposes. Seems to work reasonably well so far and seems to construct the links properly. Much more testing still to be done as well as code refactoring, etc.