Removing unneeded rgba conversion, resetting width to previous value #691
The rgba conversion added by the previous commit was not needed and could be performed with the hex values after all (issue was that the received values can be either 4 or 6 characters long).

Noticed a bug introduced in the previous commit whereby the length of the bar being 100% caused it to stick out the side of the card. Reset it back to that specific 66.5%.