Removing boolean option to disable header wrapping #722
Previously, a boolean could be passed to DataHeader to determine whether or not it should flexWrap. In the majority of cases, the header would wrap except in the case of dates when wrapping was turned off to avoid the header becoming too tall.

The option is being disabled completely to ensure consistency across the board. No headers wrap anymore