Fixing unit tests #566
A console error message was being output that wasn't failing the test but was caused by a link being generated with no text. This turned out to be expected behaviour which a test was checking for and the error is now being caught properly.

The failing unit test with removing from the cart was caused by the test data not being up to date with all IDs of search results. Fixing this and making IDs in test data consistent.