Refactoring study landing component, adding basic check for null investigation #711
This adds the beginnings of checks for a null investigation being attached to a study. If an investigation is found, it renders a LinkedInvestigation component, which contains the standard visit id, doi, instrument, release date and add to cart button. If investigation is null, it currently shows a message saying 'Missing investigation'. This may be changed later on.

Still to do more thorough checks for null across the component and address the investigations tab