Fixing problems with missing links, adding e2e tests #566
This addresses an issue where datafile links were not loading and some investigation links were missing. This was caused by a query to the datafiles endpoint missing out an include for a parent investigation, which meant the link could not be properly constructed. Adding e2e tests to ensure this is now properly tested for.