Fixing column dataKeys #566
Incorrect dataKey values for links to parent investigations/datasets were causing the page to crash whenever the user hovered over one of the links to view the popup of its full title/name. This was because the key was pointing to an investigation/dataset object as a whole and not its actual title/name. These dataKeys have been changed to the right values so that the right data is now referenced