Removing unnecessary text at bottom of page #689
This removes an instruction at the bottom of the page which instructs the user to fill in their query to the search bar as it is not deemed necessary. This was done by refactoring the searchPageTable component which displayed this text. This component now only displays if a query has been launched, which improves efficiency.