Adding 'Enter' functionality to date boxes #689
Extending the enter functionality to the date boxes as well. This allows for users to enter in the dates they wish and press enter to initiate a search, similar to how it works in the search text box. Tests have been created, moved and deleted to account for this. Also renaming searchTextBox.tsx to searchTextBox.component.tsx inline with naming standards.