Refactoring isis and dls card views, fixing filter bug, editing download button #701
ISIS and DLS card views now use react query. A bug in useFilter was preventing additional filters from being properly passed to fetchFilters, which is being fixed. Also amending the downloadButton component to take an optional variant parameter. This is because the style of the download button changes between facilities