Adding unit tests for generic datafile table, changing aggregated data function #701
Generic datafile table now has unit tests. Many are skipped with a mind to move them to a parent component later but are currently remaining to act as a reference. One is skipped but only because I haven't figured out how to make it pass yet

Also changing how aggregated data in components is compiled to be consistent with how it is done in common. This depends on whether or not the view we are in needs a different data format and whether or not the data has a pages field that we can call flat() on