Changing sort by labels to upper case so they are all read aloud properly #772
This also changes their value from the dataKey of the object to the label of the object. This is because of a jarring effect that can occur when a field is obtained through a complex relationship, e.g. an instrument's investigation. So this would lead to the aria label being 'sort by instrument.investigation'. To make this consistent with what the user sees on the screen, it now says 'sort by ${label-value}'