Unskipping my data tests using Thomas409 user #659
This changes the cypress config to make use of an existing user instead of a test user, in order to fetch data that is mapped to them. The user Thomas409 is used as they happen to have the most investigations mapped to them. With this returned data, we can now properly e2e test my data