Adding investigation and dataset card views #420
Creating investigation and dataset card view components. These currently display by default for testing purposes.

Also amending the key used for linking datasets to parent investigations. Previously this was the investigation title but this is too long to be practical for either the table or card views. Changing this to investigation name to allow for something shorter