Selections tab is now loaded first every time #1193
Previously, the last tab the user had open was displayed when /download was visited. The last tab number was persisted in session storage to facilitate this. However, because we have links pointing to "selections", it can be confusing if the user arrives at the downloads tab instead and has to then navigate to the selections tab themselves. This commit remedies this by making the selections tab the goto every time. From now on, the user can only reach the downloads tab by manually clicking on the tab in /download.