Removing seconds from date parsing, removing index.tsx changes #1175
Now completely removing the ability to set the seconds value in the admin download table date filtering. Previously, this was possible and causing issues with expected results. To align with the date picker GUI, we can now only go down to the minute level.

Also removing changes made to index.tsx. This was to mistakenly account for time inputs to be filtered but this is not needed in index as it can only receive a set date value. As such, the start date will carry on with a time of 00:00:00 and the end date will carry on with a time of 23:59:59 inclusive