Disabling download button and showing tooltip if file size equals zero #1195
This involves passing the size of the relevant entity to the download button. In table and card view, this involves making use of the sizeQueries array. If the button is disabled, the user can move their mouse over the button to show a tooltip explaining the file is empty and thus cannot be downloaded.