Customising notification message if it concerns session expiry #1209
To enable a more helpful notification message for the user, we will amend the returned error message from TopCAT which references a null or expired session ID which the user doesn't need to know or care about. Instead, the user is now informed that their session has expired and is instructed to reload the page to refresh their session.